[flue] Replace style-guide child-workflow fan-out with in-process sessions
Collapse the per-file style-guide review from N child workflows (admitted
over HTTP + Durable Streams long-poll) into concurrent sessions within the
code-review orchestrator's own Durable Object. One harness over a single
shared workspace, hydrated once, with one session per file fired via a
concurrency pool; a single file's failure degrades to an empty result
instead of aborting the whole review.

This removes the separate child-workflow DOs that were the unit getting
evicted mid-poll (the interruption that failed PR #31514), and deletes the
admit/poll plumbing (lib/poll-run.ts retained for the spam-filter gate).

- Add lib/style-guide-inproc.ts (hydrate-once + per-file session fan-out)
- Rewrite orchestrator fan-out to call runStyleGuideReviewInProcess
- Delete lib/style-guide-fanout.ts (only the orchestrator used it)
- Retire the style-guide-review workflow + its DO class (v4 deleted_classes migration)
